+++ console/beacn/__init__.py
@@ -1,0 +1,1 @@
+"""BEACN shared application package."""